Day 1 One Liners

+++ 2025/aoc.py
@@ -40,17 +40,22 @@ if __name__ == "__main__":
     now = datetime.datetime.now()
     today = sys.argv[1] if len(sys.argv) == 2 else now.day
     year = now.year
-    template = """from aoc import input_for
+    template = """def pt_1(puzzle_in: str):
+    pass
 
-def pt_1():
-    puzzle_in = input_for(day={day}, year={year})
+def pt_2(puzzle_in: str):
+    pass
 
-def pt_2():
-    puzzle_in = input_for(day={day}, year={year})
+pt_1_oneline = lambda puzzle_in: ...
+pt_2_oneline = lambda puzzle_in: ...
+
 
 if __name__ == "__main__":
-    print("part one: ", pt_1())
-    print("part two: ", pt_2())"""
+    from aoc import input_for
+
+    puzzle_in = input_for(day={day}, year={year})
+    print("part one: ", pt_1(puzzle_in), pt_1_oneline(puzzle_in))
+    print("part two: ", pt_2(puzzle_in), pt_2_oneline(puzzle_in))"""
 
     dir: pathlib.Path = pathlib.Path.cwd()
     if not dir.name.isdigit():
@@ -58,8 +63,11 @@ if __name__ == "__main__":
 
     dir.mkdir(exist_ok=True, parents=True)
     file = dir / f"day_{today:02}.py"
-
-    file.write_text(template.format(day=today, year=year))
-    print("Created a new file for today!")
+    
+    if file.exists():
+        print("A file for this day already_exists.")
+    else:
+        file.write_text(template.format(day=today, year=year))
+        print("Created a new file for today!")
 
     input_for(int(today), year)
